revisitation de l'inventaire et création Forgeron

+++ classe.go
@@ -3,7 +3,7 @@ package personnage
 type Kryptonien struct {
 	Viemax       int
 	Inventaire   int
-	Puissance    int
+	Puissance    float64
 	Faiblaisse   []string
 	Agilite      int
 	Intelligence int
@@ -18,7 +18,7 @@ var Kryptonien1 = Kryptonien{20, 10, 25, []string{"Crypto,", "magie,", "puissanc
 type Batfamily struct {
 	Viemax       int
 	Inventaire   int
-	Puissance    int
+	Puissance    float64
 	Agilite      int
 	Intelligence int
 	CapacityDisplay     []string
@@ -29,7 +29,7 @@ var BatFamily1 = Batfamily{10, 15, 10, 35, 50, []string{"Adaptation,", "Gadjet,"
 type Hulk struct {
 	Viemax       int
 	Inventaire   int
-	Puissance    int
+	Puissance    float64
 	Faiblaisse   []string
 	Agilite      int
 	Intelligence int
@@ -41,7 +41,7 @@ var Hulk1 = Hulk{30, 4, 50, []string{"Hypnoses,", "Self Controle,"}, 15, 10, []s
 type Kantin struct {
 	Viemax       int
 	Inventaire   int
-	Puissance    int
+	Puissance    float64
 	Agilite      int
 	Intelligence int
 	CapacityDisplay     []string
@@ -52,7 +52,7 @@ var Kantin1 = Kantin{120, 60, 100, 70, 100, []string{"Toutes,", "HO,", "Prout Br
 type Wakandais struct {
 	Viemax       		int
 	Inventaire   		int
-	Puissance    		int
+	Puissance    		float64
 	Faiblaisse   		[]string
 	Agilite      		int
 	Intelligence 		int
